filler chameau

+++ libft/srcs/ft_putstr.c
@@ -3,16 +3,17 @@
 /*                                                        :::      ::::::::   */
 /*   ft_putstr.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: ahadjer <marvin@42.fr>                     +#+  +:+       +#+        */
+/*   By: ahadjer <ahadjer@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/03/20 09:18:17 by ahadjer           #+#    #+#             */
-/*   Updated: 2017/03/20 09:18:18 by ahadjer          ###   ########.fr       */
+/*   Updated: 2018/07/12 18:31:55 by lchety           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
+#include <stdio.h>
 
 void	ft_putstr(char const *s)
 {
-	write(1, s, ft_strlen(s));
+	write(1, s, (int)ft_strlen(s));
 }
